Split `queries.py` into a `queries/` package
The single 2400-line module was really five: shared plumbing, connection
lifecycle, and one file per resource (sessions, nodes, LLM traces, events,
middleware, guardrails). Splitting each into its own file makes the
boundaries load-bearing instead of aspirational — a change to the
middleware pipeline touches middleware.py and doesn't scroll past LLM
trace logic.

Public surface is unchanged: `routes.py` still does `from . import
queries` and every callable it uses is re-exported from
`queries/__init__.py`. Verified end-to-end against the real event stream
— every endpoint returns 200.

- `_common.py`: `_rows`, `_in_clause`, `_window_predicates`,
  `_order_clause`, `_parse_json`, and the shared denormalizing CTEs
  (session / node / node-span / middleware-name / llm-creation).
- `_connection.py`: the shared `EventQuery` singleton and its refresh.
- `sessions.py`, `nodes.py`, `llm_traces.py`, `events.py`,
  `middleware.py`, `guardrails.py`: one resource each.
